Sannsynlighetsfordeling plottet.

+++ FYS3150/Oblig4/ising.h
@@ -4,12 +4,13 @@
 #include <random>
 #include <cstdlib>
 #include <cstring>
+#include <cmath>
 
 int index(int i, int L);
 
 int calculate_energy(int **spins, int L);
 
-void output(FILE* file, int i, double T, int E, int E2, int M, int M2, int Mabs);
+void output(FILE* file, int i, double T, double E, double E2, double M, double M2, double Mabs);
 
 void ising(FILE* file, const char* mode, int L, int N, int dN, double T);
 #endif
